+++ lib/routes/size.dart
@@ -1,15 +1,17 @@
 import 'package:flutter/material.dart';
 
-class Sizing {
-  static fullW(BuildContext context) => MediaQuery.of(context).size.width;
-  static halfW(BuildContext context) => MediaQuery.of(context).size.width / 1.5;
-  static customW(BuildContext context, int num) =>
+/*
+FSZ is Floxy Size
+
+*/
+class FSZ {
+  static double fullW(BuildContext context) => MediaQuery.of(context).size.width;
+  static double halfW(BuildContext context) => MediaQuery.of(context).size.width / 1.5;
+  static double customW(BuildContext context, int num) =>
       MediaQuery.of(context).size.width / num;
-  static fullH(BuildContext context) => MediaQuery.of(context).size.height;
-  static halfH(BuildContext context) =>
+  static double fullH(BuildContext context) => MediaQuery.of(context).size.height;
+  static double halfH(BuildContext context) =>
       MediaQuery.of(context).size.height / 1.5;
-  static customH(BuildContext context, int num) =>
+  static double customH(BuildContext context, int num) =>
       MediaQuery.of(context).size.height / num;
 }
-
-get sizing => Sizing();